+++ commands/sync_test.go
@@ -19,32 +19,33 @@ func TestSync(t *testing.T) {
 	expectedMachine := "mymachine"
 	expectedUser := "myuser"
 	os.Args = []string{"", "create", "-m", expectedMachine, "-u", expectedUser}
-	buf := new(bytes.Buffer)
+	inBuf := new(bytes.Buffer)
+	outBuf := new(bytes.Buffer)
 	os.Remove("qa/passwords.db")
-	actualRet := Main(buf)
+	actualRet := Main(inBuf, outBuf)
 	expectedRet := 0
 	if actualRet != expectedRet {
-		t.Fatalf("Main(create) = %v, want %v, output is %q", actualRet, expectedRet, buf.String())
+		t.Fatalf("Main(create) = %v, want %v, output is %q", actualRet, expectedRet, outBuf.String())
 	}
 	os.Rename("qa/passwords.db", "qa/remote.db")
 	os.Args = []string{"", "sync"}
-	buf = new(bytes.Buffer)
+	outBuf = new(bytes.Buffer)
 
-	actualRet = Main(buf)
+	actualRet = Main(inBuf, outBuf)
 
 	expectedRet = 0
 	if actualRet != expectedRet {
-		t.Fatalf("Main(sync) = %v, want %v, output is %q", actualRet, expectedRet, buf.String())
+		t.Fatalf("Main(sync) = %v, want %v, output is %q", actualRet, expectedRet, outBuf.String())
 	}
 	os.Args = []string{"", "search", "-m", expectedMachine, "-u", expectedUser}
-	buf = new(bytes.Buffer)
-	actualRet = Main(buf)
+	outBuf = new(bytes.Buffer)
+	actualRet = Main(inBuf, outBuf)
 	expectedRet = 0
 	if actualRet != expectedRet {
 		t.Fatalf("Main(search) = %q, want %q", actualRet, expectedRet)
 	}
 	expectedOutput := "machine: mymachine, service: http, user: myuser, password type: plain, password: output-from-pwgen\n"
-	actualOutput := buf.String()
+	actualOutput := outBuf.String()
 	if actualOutput != expectedOutput {
 		t.Fatalf("actualOutput = %q, want %q", actualOutput, expectedOutput)
 	}